requirements.txt
streamlit==1.12.0
tensorflow==2.18.0
numpy==1.24.0
Pillow==9.4.0
scikit-learn==1.2.2
h5py==3.8.0
pickle5==0.0.11
